Add deserialization tests for remaining DispatchEvents
